chore: move ui config page to comp
